Rearranged functions into multiple files
