+++ xsd_validator/src/error.rs
@@ -1,0 +1,8 @@
+#[derive(Debug)]
+pub enum XSDValidationError {
+    PathError,
+    StringError,
+    OutputRedirectError(String),
+    GenerateContextError,
+    ParseError(String),
+}
